Changed shell executions to use non-deprecated methods
Currently, both the PowerShell and CMD shell executions are using deprecated APIs
to do so. As a result, each time you use this plugin, a warning is logged that
you are using a deprecated API. I have changed the plugin to use the new,
non-deprecated APIs for both PowerShell and CMD execution that were introduced
in WinRM 1.5.0.